+++ RELEASE_NOTES.md
@@ -5,6 +5,9 @@ v0.23.0 - 9/07/2026
 - Converted some tools to kwiver-style applets
 
 
+- Add viame json tool for DIVE and COCO json filtering and statistics
+
+
 - UI Updates for multi-camera (EO/IR/UV) registration, droped frames, and joint panning